requirements.txt
openai-whisper
pydub
yt-dlp
PyDrive